Improved security by moving OTP generation logic to native code.
The provided C code only generate 20 bytes, that needs to be processed in a  way to output a six digits number. The remaining code can be implemented in both Java or C. Unlike Java, C code cannot be decompiled only disassembled. This make reverse engineering harder when compared to Java, which even with obfuscation can be read after disassembled, leaving intellectual property vulnerable. In this application no one needs to know how the six digits number is generated. That's why it was chosen to move this code to native code.